[Port] [6000.4] Remove usage of deprecated instanceID APIs in com.unity.render-pipelines.core

+++ Packages/com.unity.render-pipelines.core/Runtime/RenderGraph/RenderGraphResourceBuffer.cs
@@ -171,9 +171,9 @@ namespace UnityEngine.Rendering.RenderGraphModule
             return "GraphicsBuffer";
         }
 
-        override protected int GetSortIndex(GraphicsBuffer res)
+        override protected ulong GetSortIndex(GraphicsBuffer res)
         {
-            return res.GetHashCode();
+            return (ulong)res.GetHashCode();
         }
     }
 }
